Research Projects/Pinn POC/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Scientific computing
scikit-learn>=1.0.0
scikit-image>=0.19.0

# Data handling
h5py>=3.6.0
tqdm>=4.62.0

# Configuration and logging
pyyaml>=6.0
omegaconf>=2.1.0

# Jupyter and documentation
jupyter>=1.0.0
jupyterlab>=3.0.0
notebook>=6.4.0
nbconvert>=6.4.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991

# Optional GPU acceleration
# Uncomment if CUDA is available
# torch-audio>=0.13.0
# cupy-cuda11x>=11.0.0  # Adjust CUDA version as needed

# Edge deployment (optional)
# onnx>=1.12.0
# onnxruntime>=1.12.0
# tensorrt>=8.4.0  # For NVIDIA edge devices